+++ benches/outfit_gauss_iod.rs
@@ -6,14 +6,15 @@ mod benches_impl {
 
     use camino::Utf8Path;
     use criterion::{black_box, BatchSize, Criterion};
-    use outfit::constants::{ObjectNumber, TrajectorySet};
+    use outfit::constants::ObjectNumber;
     use outfit::error_models::ErrorModel;
     use outfit::initial_orbit_determination::gauss_result::GaussResult;
     use outfit::initial_orbit_determination::IODParams;
     use outfit::observations::observations_ext::ObservationIOD;
-    use outfit::observations::trajectory_ext::TrajectoryExt;
     use outfit::outfit::Outfit;
     use outfit::outfit_errors::OutfitError;
+    use outfit::trajectories::trajectory_file::TrajectoryFile;
+    use outfit::TrajectorySet;
     use rand::rngs::StdRng;
     use rand::SeedableRng;